spatialDEG

spatialDEG to perform differentially expressed gene analysis by leveraging spatial information.

spatialDEG_test(
  spa1,
  spa2,
  W,
  Y,
  Initial_theta,
  num_ls = 20L,
  mu_fixed = TRUE,
  verbose = TRUE
)

Arguments

spa1
spatial location for the first dataset.

spa2
spatial location for the second dataset.

W
covariates for the two datasets

Y
stacked gene expression

Initial_theta
initial value of theta

num_ls
total number of length scale and periodicity

mu_fixed
logical value specifying whether mu is fixed

verbose
logical value specifying whether to print the parameters during the spatialDEG analysis or not

Value

List of model parameters

Details

spatialDEG_test run spatialDEG model.

Author

Yi Yang